3 ms (33.78%) 7.9 MB (8.11%) 1318-minimum-flips-to-make-a-or-b-equal-to-c 05-09-2025 Medium .c

+++ 1318-minimum-flips-to-make-a-or-b-equal-to-c/1318-minimum-flips-to-make-a-or-b-equal-to-c.c
@@ -1,9 +1,9 @@
 
 
 int minFlips(int a, int b, int c){
-    int result=0;
-    int bitCount=(int)floor(log2((a>b)?((a>c)?a:c):((b>c)?b:c)))+1;
-    for(int i=0;i<bitCount;i++){
+    uint8_t result=0;
+    uint8_t bitCount=(int)floor(log2((a>b)?((a>c)?a:c):((b>c)?b:c)))+1;
+    for(uint8_t i=0;i<bitCount;i++){
         if((c>>i)&1){
             if(!(((a>>i)&1) ||((b>>i)&1))){
                 result++;
